fix: dragon won't try finding target when ridden

+++ src/main/java/net/dragonmounts/event/ClientMisc.java
@@ -84,7 +84,7 @@ public class ClientMisc {
             final float MAX_GLOW_BRIGHTNESS = 1.0F;
             final long NANO_SEC_PER_SEC = 1000L * 1000L * 1000L;
             long cyclePosition = System.nanoTime() % (GLOW_CYCLE_PERIOD_SECONDS * NANO_SEC_PER_SEC);
-            double cyclePosRadians = 2 * Math.PI * cyclePosition / (double) (GLOW_CYCLE_PERIOD_SECONDS * NANO_SEC_PER_SEC);
+            double cyclePosRadians = 2 * Math.PI * cyclePosition / (GLOW_CYCLE_PERIOD_SECONDS * NANO_SEC_PER_SEC);
             final float BRIGHTNESS_MIDPOINT = (MIN_GLOW_BRIGHTNESS + MAX_GLOW_BRIGHTNESS) / 2.0F;
             final float BRIGHTNESS_AMPLITUDE = (MAX_GLOW_BRIGHTNESS - BRIGHTNESS_MIDPOINT);
             int brightness = MathHelper.clamp((int) (255 * (BRIGHTNESS_MIDPOINT + BRIGHTNESS_AMPLITUDE * MathHelper.sin((float) cyclePosRadians))), 0, 255);
